+++ zeldris/modules/sql/antiflood_sql.py
@@ -18,7 +18,7 @@
 import threading
 
 from sqlalchemy import String, Column, Integer, UnicodeText
-
+from sqlalchemy.sql.sqltypes import BigInteger
 from zeldris.modules.sql import SESSION, BASE
 
 DEF_COUNT = 0
@@ -29,7 +29,7 @@ DEF_OBJ = (None, DEF_COUNT, DEF_LIMIT)
 class FloodControl(BASE):
     __tablename__ = "antiflood"
     chat_id = Column(String(14), primary_key=True)
-    user_id = Column(Integer)
+    user_id = Column(BigInteger)
     count = Column(Integer, default=DEF_COUNT)
     limit = Column(Integer, default=DEF_LIMIT)
