+++ zeldris/modules/sql/userinfo_sql.py
@@ -17,14 +17,14 @@
 
 import threading
 
-from sqlalchemy import Column, Integer, UnicodeText
-
+from sqlalchemy import Column, UnicodeText
+from sqlalchemy.sql.sqltypes import BigInteger
 from zeldris.modules.sql import SESSION, BASE
 
 
 class UserInfo(BASE):
     __tablename__ = "userinfo"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     info = Column(UnicodeText)
 
     def __init__(self, user_id, info):
@@ -37,7 +37,7 @@ class UserInfo(BASE):
 
 class UserBio(BASE):
     __tablename__ = "userbio"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     bio = Column(UnicodeText)
 
     def __init__(self, user_id, bio):
